Rewrite parser and compiler to use AST
Instead of compiling to bytecode in a single pass, restructure the
parser to build an AST and then talk the AST to compile bytecode. This
makes the code much easier to follow, and also makes it much easier to
tell when part of the code should be treated as a parser, value, or
pattern. This is because we now compile top-down from the root node of
the AST, instead of bottom-up from the first expression in the input.

Also update StringTable so that retrieving a string that does not exist
is a runtime panic. This simplifies a lot of code that accesses interned
strings. If we ever try to access a string that doesn't exist that would
be a bug.