Restructure Backtrack and Or ops to be tail call eligible
- Introduce `inputMarks` stack to the VM. This stack tracks positions in
  the input which might be needed later in parsing.
- Rework the Backtrack op code to push an input mark before executing
  its left-hand-side parser, and then jump back to that position if the
  lhs succeeds, before parsing the rhs.
- Rework the Or op code to push an input mark before executing its
  left-hand-side parser, and then jump back to that position if the lhs
  fails, before parsing the rhs.
- Remove the `start` and `end` fields from the `ParseResult.Success`
  struct. These fields were used to track the input segment
  corresponding to the parsed value, but that data was only needed in
  order to reset the input position when required by Backtrack and
  Or.

This change makes functions called on the rhs of Or eligible for tail
call elimination, because the cleanup work which was done after running
the rhs parser can now be done before the rhs parser, leaving it in tail
position. The rhs of Backtrack was already a tail position, but now uses
the inputMarks stack. In the following commit we'll extend this change
to make other infix ops tail call elimination eligible.