Restructure TakeRight and Conditional ops to be tail call eligible
When compiling the TakeRight and Conditional infix ops, emit the
bytecode for the rhs parser as the last step, so that if a function is
invoked it will be in tail call position.

In the case of TakeRight parse the lhs and then discard the result. If
the result was a failure then jump ahead instead of parsing the rhs. If
it was a success then continue on and parse the rhs. This prevents the
parsed stack from growing, while also avoiding stack manipulations after
parsing the rhs. This is possible because we no longer need to use the
success values in the parsed stack to track input position.

In the case of conditionals, ConditionalThen discards the `if` parser
and jumps if the result was a failure, so that the `then` parser is
skipped over. ConditionalElse is skipped over unless the `if` parser was
successful and the `then` parser was ran, in which case it skips over
the rhs else parser.