Try to resolve variable in pattern before destructuring
Handle the case where the left hand side pattern is a variable, and that
variable is already bound to a local or global value. Fetch the value,
push it on to the stack, and then compare the parsed result to the
value. If the variable is unbound this will result in a runtime error.