Clear stack on program end
Remove a special case that would clear the stack on function end
_unless_ the current function was the main function. Also remove an
assertion that there's one item on the stack after running the VM.

In the next commit we're going to introduce locals used in pattern
matching, which means the main function can have locals left on the
stack. This means the stack size assertion will not always hold. We
could keep the special case for not clearing the stack for the final
call frame, but it doesn't really matter so let's cut it.