Add new locals used in pattern matching
Before emitting bytecode to perform pattern matching, walk the AST for
the pattern and find all variables. If the variable is already
introduced in the local scope (either from a function param or a
previous pattern match) then do nothing. If the variable is not yet in
global scope then add it, and then check to see if it's shadowing a
global value with the same name. If so copy the global value into the
local variable slot.

With this change we can use pattern matching to compare a parsed value
to a pattern variable bound in the local scope. If the variable remains
unbound then we crash, since assigning via pattern matching is not yet
implemented.