Fix parsing failure when function arg is an anonymous function
The binding power of `,` was incorrectly defined in the parser. This
means that the list of args/params put infix operators above
`FunctionArgOrParam` in the parse tree, creating a malformed AST. To fix
this we lower the binding power so that complex args are grouped
together in the AST.

Add a test to confirm the order of params doesn't get flipped -- this
would happen if we made `FunctionArgOrParam` right associative instead
of left associative.