Don't use parsePrecedence for args/params list
Parsing function args/params list used a cute trick where the comma was
considered as an infix operator and the fill list was parsed as one tree
using `parsePrecedence`. This produced bad error messages whenever a
program used a comma in an unusual place. Instead we iterate through
comma separated lists in a specialized recursive function.