Allow function calls to have both parser and value arguments
When compiling a function call try to figure out if arguments to the
function should be parsers or values, and then compile each arg
accordingly. This method only works if the callee function is global,
since we don't know enough at compile time about parsers passed as args
into a function. For now skip this case.

Add compile time checks that global functions are called with the
correct number of arguments.

Enable the tail recursive version of the `many` and `until` stdlib
parsers, also enable a number of array parsers.