Replace @Success values in final output with empty strings
The final output should be valid JSON, so we need to replace internal
values like @Success with some default. By waiting until the very end we
can safely replace each @Success with an empty string, since if @Success
needed to act as a different kind of empty value that would have
happened by this point.