Implement pattern matching on objects with bound variables as keys
Paterns can include variables in place of object keys. If the variable
is bound then the variable value will be substituted in the pattern.

The following will succeed with `1` because `Foo` is bound:
```
Foo = "foo"

const({"foo": 1}) -> {Foo: N} $ N
```

The following will throw a runtime error because `Foo` is unbound:
```
const({"foo": 1}) -> {Foo: 1}
```

Eventually this second case will be supported by having the pattern
match find the first available value that matches `1` and then bind the
corresponding key, which would assign `"foo"` to `Foo`.