track function locals in the compiler instead of the Function dyn elem
Attempt to make Function smaller by moving out info that's only needed
at compile time. The issue is that we need to track the types of
function params across modules, so if I'm compiling a function call for
a function from a previously compiled module I still have the param
info.